advanced-rag-backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0

sqlalchemy==2.0.29
alembic==1.13.1
psycopg2-binary==2.9.9   # if using Postgres later
python-multipart==0.0.9

passlib==1.7.4
python-jose==3.3.0
bcrypt==4.1.2

pydantic==1.10.14

# FAISS for Windows (CPU-only)
faiss-cpu==1.7.4

# Embeddings (small footprint)
sentence-transformers==2.5.1

# Groq LLM
groq==0.5.0

# File parsing
python-docx==1.1.0
pdfplumber==0.11.0

# Utils
python-dotenv==1.0.1
loguru==0.7.2
